requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
jsonschema==4.23.0
python-multipart==0.0.9
python-dotenv==1.0.0
tenacity==8.2.3
python-docx==1.1.2
pypdf2==3.0.1
anthropic>=0.40.0
